Ensure flowindent tracking is done across all events in a buffer
The flowindent tracking datastructure was local to the processing of an
event (do_consume_one()) which is pretty useless because it is meant to
keep state across multiple events in a buffer.  It has been moved to
do_consume_cpu() where it belongs, and is now passed to dt_consume_one().

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>