Fix creation of FBT return probes
The FBT provider was looking for a EVENTSFS/kprobes/<fun>/format to
determine the event id for the probe.  If this file did not exist, a
new kprobe or kretprobe (entry vs return) would be created and we
would try to read that file again.  Eugene Loh flagged the problem
in the implementation: both entry and return probe for a given
function were retrieving the id for the same kprobe event, as given
by the file pattern shown above.

Since a given function can (and often will) have an entry probe and
a return probe associated with it, they must be created as distinct
events in the kernel tracing infrastructure.

Dynamically created probes like kprobes, kretprobes, and uprobes
(used for the BEGIN and END probes) are grouped based on the dtrace
instance that creates them (pid), provider name, and an optional
extra distinction.  The BEGIN and END probes are created under the
dt_<pid>_dtrace group, whereas FBT entry probes are created under
the dt_<pid>_fbt_entry group, etc.

The dtrace provider probe creation and removal have been updated in
favour of a more generic implementation that is shared with the FBT
provider.

The SDT provider has been updated to filter out any probes from our
groups.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>