Fix various memory leaks
Testing dtrace with valgrind identified a bunch of significant memory
leaks.  These have been fixed.

There is still some loss of memory due to string elements in probe
descriptions not consistently using strdup() to ensure that they can
be freed when no longer needed, and also in the buffer handling in
the lexer.  These are left for future investigation and resolution.

Orabug: 31386368
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>