Provide a more consistent implementation of the trace() action
The trace(n) action in DTrace v1 does not provide a consistent way
for displaying numeric data.  Due to the fact that information about
the value is lost (whether it is a signed value), the consumer was
always interpreting the value as a signed value.  However, due to
what seems like a bug, values with an integer datatype of less than
64 bits were being interpreted as 32-bit values.  Therefore;

                                  (quiet)   (non-quiet)
	(int8_t)-1	prints as 255       " 255"
	(int16_t)-1	prints as 65535     " 65535"
	(int32_t)-1	prints as -1        "       -1"
	(int64_t)-1	prints as -1        "               -1"

        (uint8_t)0xff   prints as 255       " 255"
        (uint16_t)0xffff
                        prints as 65535     " 65535"
        (uint32_t)0xffffffff
	                prints as -1        "       -1"
        (uint32_t)0x7fffffff
	                prints as 2147483647
                                            " 2147483647"
        (uint64_t)0xffffffffffffffff
                        prints as -1        "               -1"
        (uint64_t)0x7fffffffffffffff
                        prints as 9223372036854775807
                                            " 9223372036854775807"

The output statements fail to specify the correct number of decimal
characters in the output for 32-bit and 64-bit values.

This patch provides a consistent implementation by passing a flag
into the record to indicate whether the value is signed.  The output
uses a custom formatting string based on the datatype width and sign
of the value.

With the new implementation:
                                  (quiet)   (non-quiet)
        (int8_t)-1      prints as -1        "  -1"
        (int16_t)-1     prints as -1        "    -1"
        (int32_t)-1     prints as -1        "         -1"
        (int64_t)-1     prints as -1        "                   -1"

        (uint8_t)0xff   prints as 255       " 255"
        (uint16_t)0xffff
                        prints as 65535     " 65535"
        (uint32_t)0xffffffff
                        prints as 4294967295
                                            " 4294967295"
        (uint64_t)0xffffffffffffffff
                        prints as 18446744073709551615
                                            " 18446744073709551615"

This patch also adds unit tests for the various numeric value sizes,
and a variety of error conditions.

Orabug: 31407534
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>