First implementation of the printf action
The legacy implementation was a close interaction between the userspace
and kernel components of DTrace.  Now that everything is in userspace
significant portions of the printf-style actions are updated.

Format strings used to be converted into a data structure that describes
the format and its conversions.  This representation was used during the
compilation of clauses to validate arguments.  Then, the string was
passed to the kernel component of DTrace, to be retrieved when buffer
processing needed to format data.  It would request the format string
based on a numeric id, and then process the returned string into a data
structure (almsot identical to the one used during compilation).

In the new version, we can simply convert the format string into the
final data structure and store that in dtrace_recdesc_t.  A reference
to the data structure (dt_pfargv_t *) is passed to dt_cg_store_val()
which in turn passes it to dt_rec_add() as format.  This ensures that
when output records are processed, we know that there data items belong
to a printf-style action.

A bug existed in the error reporting for multiple '*' specifiers in
the field width.  It mistakenly reported it as "more than one '*'
specified for the output precision" whereas it should refer to the
"output width".

Note that only numeric data is supported at this point due to fact that
string support is still missing in DTrace v2.

A series of unittests are included in test/unittest/actions/printf to
ensure the correct operation of the printf-style actions.

Orabug: 31236012
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>