tests: fix tst.errno2.d to work with openat and skip aarch64
tst.errno2.d previously was tagged unstable, because there were failures
due to using this->fn instead of self->fn.  This commit changes to self->fn
and removes the unstable tag.

Another problem solved is the test was written to only work with the 'open'
function call and some systems now have 'openat' instead.  The test now
works with both functions.

The test has not worked with the aarch64 implementation, because the test
makes use of copyinstr() which is unavailable on aarch64.  This commit
provides for skipping tst.errno2.d on aarch64 via the addition of the file
tst.errno2.aarch64.x

A minor modification has also been made to tst.errno2.r to match the
modified output.

Signed-off-by: David Mc Lean <david.mclean@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>