drti: reference first loaded segment of object containing dtrace_dof_init
Any ELF executable or shared library (object) that contains USDT probes
contains a DOF section which describes the probes and also contains a copy
of the dtrace_dof_init() function.  This dtrace_dof_init() function is used
to register the DOF data of the object with the kernel.

The dtrace_dof_init() function needs to identify what type of object it is
being invoked for because the address representation varies (for ELF object
file type ET_DYN the symbol values in the DOF are relative to the load
address of the object rather than absolute values).  The ELF object file
type is obtained from the ELF header and the ELF header is found in the
first mapping (or segment) of the object.

The implementation used previous to this commit was to loop through all
mappings looking for one that contains the address of the dtrace_dof_init()
function.  The assumption was that the executable code mapping with the
dtrace_dof_init() function would always be the first mapping, and since the
ELF header is in the first mapping, the mapping with the dtrace_dof_init()
function would also contain the ELF header.

The assumption was proven wrong where sometimes the dtrace_dof_init()
function is not in the first mapping, and this commit updates the code to:
First loop over the mappings to determine the name of the object containing
the executing dtrace_dof_init() instance and record this object name.
Then a second loop is run over the mappings to determine the mapping with
the lowest start address (which is in fact the first mapping whose name
matches the one recorded from the first loop).

This first mapping of the object contains the ELF header (this ELF header is
used to determine the ELF object file type) and the first mapping start
address is the base address to be passed to the kernel along with the DOF
data.

Signed-off-by: David  Mc Lean <david.mclean@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>