Implement the curcpu builtin D variable
The 'curcpu' builtin D variable is special in the sense that it is
a pointer to a cpuinfo_t struct for the CPU on which the D program is
executing, yet no such structure exists in the kernel (DTrace v1 used
to provide it through a kernel patch).  In order to provide the same
functionality, a new BPF map 'cpuinfo' is populated from userspace
with cpuinfo_t structs for each CPU.

The BPF map creation code had an entry for a 'probes' map that is no
longer needed, and it has therefore been removed (incl. the probec
argument to dt_bpf_gmap_create().)

To facilitate using map FDs at time of creation, create_gmap() now
returns the fd.

To satisfy the BPF verifier (and because it is good practice), the
code generator will now ensure that a pointer lvalue is not NULL prior
to accessing its members.

The builtin D variables currently supoorted in dt_get_bvar() have also
been reordered to reflevt the order of their IDs (for consistency).

Orabug: 31221984
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>