Allow undefined references to BPF maps in precompiled functions
Up until now, BPF maps needed to be declared in C code and encoded in
the 'maps' section of the ELF object of compiled BPF code.  Upon
loading the ELF object, the maps mentioned in the 'maps' section were
matched against the maps DTrace creates (and if not found, the map
would be registered as a new map).  Since DTrace is creating the maps
prior to loading any BPF code, it is sufficient to have references to
the maps as undefined symbols in the ELD object.

The new code supports both forms of specifying maps.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>