Integrate the predicate (if any) into the main program.
The predicate was generated as a function, called from the provider
trampoline.  This poses a problem when the DTrace BPF context needs
to be accessed within the predicate code (e.g. when referencing a
builtin variable).

The new implementation integrates the predicate (if any) into the
actual main program.

Orabug: 31414309
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>