Report missing BPF dlib symbols as a proper compiler error
If a BPF dlib symbol was not found in the bpf_dlib.o object, a SEGV
would occur because the compiler would try to dereference a NULL
pointer.  This patch adds proper error handling for this case and
also fixes a few error handling cases where a goto was done to the
wrong label.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>