Do not register handlers when not executing probe clauses
The various handlers that are used by consumers only apply to the
execution of probe clauses.  For operating modes that do not execute
clauses there is no need to register handlers.  Even worse, if e.g.
the error handler is being registered with the default error clause
and -xnolibs is used (which is common with the -h option to generate
a header file), an error will occur because bpf_get_bvar() cannot be
found.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>