Clean-up for the prologue and epilogue functions
While all BPF functions must return a value, we were not making sure
that %r0 had a value (which would cause the BPF verifier to complain)
or that the value was not accidental (last value assigned to %r0).  We
now ensure we always return 0.

The prologue was generating code equivalent to (in C):

	if (!(cond))
		goto lbl;
	return;
    lbl:

Since we already have a label set at the return at the end of the
function, we can use the following instead:

	if (cond)
		goto exit;

This saves us several labels and 'exit' instructions.  In view of the
change to always return 0, the code at the exit label becomes (BPF):

	mov %r0, 0
	exit

This patch also fixes comments in the code to ensure they reflect the
way the code is supposed to work:

 - consistently talk about 'return' rather than 'exit'
 - rename the 'mem' variable to 'buf' to avoid a conflict with the
   'mem' BPF map
 - reflect the code changes mentioned in this patch
 - make the epilogue comment block consistent with the one for the
   prologue (and continue the numbering since the epilogue continues
   the function that the prologue starts)

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>