The printf action without conversions was not getting printed
Eugene Loh reported an issue with the printf action:

> Consider
>
>          # dtrace -n 'BEGIN
>          {
>              printf("Hello world.\n");
>              printf("Hello world. %d\n", 0);
>              exit(0);
>          }'
>
> The first printf() is not printed.  The second one is.
>
> I think the problem is that user space prints output based on
> "records" that are formed during BPF code generation, but the
> first printf() doesn't need any BPF code and therefore has no
> "record."

Indeed, no record descriptor was being generated for the case where
a constant format string is provided without conversions.  That
means that when the buffer is being consumed there is nothing to
indicate that output should be provided.

If printf() does not have arguments beyond the format string, we
add a record descriptor without data (kind = DTRACEACT_PRINTF,
size = 0, alignment = 1).  This ensures that when this record
descriptor is encountered in the metadata, the constant format
string will be printed.

Four tests are added to exercise this feature.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>