BPF identifiers should have DIFO as data rather than dt_bpf_func
The mechanism for managing BPF symbols loaded from the precompiled
function library was storing a dt_bpf_func struct as data member for
identifers in the BPF symbol idhash.  However, once the data has been
loaded from the precompiled function library, there is no further
need for the dt_bpf_func struct data.  All that is needed is the DIFO.

We now store a reference to the identifier in the dt_bpf_func struct
so that we can set the data member at the end of the get_symbols()
function to be the DIFO for the loaded external function.  We then
clean up the dt_bpf_func structs.

In order to make this work a tiny change was made to the idhash
destruction code because it may process identifiers that do not have
idops associated with them.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>