tests: section-oops: finally fix the -EBUSY error
This test triggers -EBUSYs at runtest termination when the tmpdir is
removed, because we fail to umount the temporary ext4 filesystem, even
though we umount it in an exit trap.

The cause is that pound_disk, which writes to the loopback fs, is
backgrounded, thus asynchronous, and mostly blocked on huge disk writes:
so there's absolutely no guarantee that it'll respond to the SIGKILL
before the exit 0 is hit and the exit trap does a umount.

So guard against this in multiple ways:

 - wait(1) for the background process to terminate after we kill it
 - don't disown: this stopped 'Killed:' messages from bash from
   disrupting .r comparison, but this particular test has no .r, so is
   not impacted, and we need bash to own this process before we can wait
   for it
 - just in case (belt-and-braces), umount lazily.  We can still rm the
   loopback image even if the lazy umount hasn't detached everything
   yet, and the end result a second or two later is just what we want:
   everything is terminated, no error messages, and all space
   relinquished.

Ever so many thanks to David for tracking this down, after I failed to
do so five times over and had given up hope of ever figuring out the
cause.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Diagnosed-by: David Mc Lean <david.mclean@oracle.com>
Reviewed-by: David Mc Lean <david.mclean@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>