The stackdepth/tst.value.r.p could loop forever
When a test has a post-processor script associated with it, this post
processor is executed after the test execution completes.  It is meant
to post-process the test output, prior to comparison with expected
output.  This means that the post-processor is executed outside the
test timeout protection mechanism, and if the post-processor contains
a bug that causes it to never complete, the testsuite execution will
hang indefinitely.

The stackdepth/tst.value.r.p post-processor was not safe in this regard
because it was using getline to read from input without checking that
there was input to read (no EOF detection) or that an error might have
occurred trying to read data.  This could lead to an endless loop.  The
script now verifies the return value of getline to ensure it successfully
read a line from input.  If not, it will report failure.

This patch also adds newline characters to the printf statements in the
actual test.  While the stack() action ensures its output starts with a
newline, and that every line in the output terminates with a newline,
it is still a good idea to ensure marker strings that are looked for in
the post-processor are guaranteed to be on their own line.

Incidentally, this issue was found because DTrace v2 currently does not
have a working stack() implementation and therefore no output is
generated for this action.  This resulted in the stack start and end
marker strings to be in the same line.  The post-processor did not
allow for that.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>