Avoid out-of-memory error report when linking BPF code without relocs
When there are no BPF relocation records, an allocation of size 0 was
being attempted.  Since that resulted in a NULL return value, it was
assumed that the allocation failure indicated an out-of-memory
condition.

We now simply do not try to allocate a BPF relocation table when there
are no BPF relocation records to be copied into it.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>