Fix double-free on printf format structure
The printf format structure was being recorded for every data record
item corresponding to a width, precision, or conversion in the format
string.  However, only the first occurrence actually gets used.  When
the data records are being freed, multiple records for the same printf
would cause the printf format structure to be freed more than once.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>