Fix NULL pointer dereference in dt_program_create()
If the allocation of a new program fails, a EDT_NOMEM error condition
is set.  But the function continues and then tries to assign the DOF
version.  But since the program is NULL at this point, that assignment
triggers a NULL pointer dereference.

The new behaviour returns NULL right away after setting the error.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>