Fix test/unittest/providers/tst.dtrace_cleanup.sh
The test would launch DTrace with two probes, BEGIN and
END, and watch for "matched [1-9] probe" in its output
to detect completion of job launch.  The problem is that
we used "dtrace -n BEGIN -n END", which emits the messages:
    dtrace: description 'BEGIN' matched 1 probe
    dtrace: description 'END' matched 1 probe
That means it's possible for us to believe both probes had
been created when only one had.  The test would occasionally
fail.

Change the test to use "dtrace -n BEGIN,END".  That way, there
would be only one message:
    dtrace: description 'BEGIN,END' matched 2 probes
and detection of the message would be a better indicator.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>