Try an orderly shutdown if a test times out
The test suite has a time-out mechanism.  Currently, it sends a
KILL signal, which does not allow DTrace to clean up after itself.

Use TERM for the time-out signal, since it (and INT) are captured
by DTrace and allow an orderly shutdown.  Specify as well that a
KILL signal should be sent if the command is still running a few
seconds after the initial signal.

The timeout return code is different depending on whether the first
or the KILL signal terminated the job, but runtest.sh already checks
this.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>