Generate trampoline programs outside of clause compilation
An earlier patch ("d58b32d6 Generate clause predicates as functions")
changed predicates from being independent actions to being compiled as
BPF functions.

This was followed by a patch ("7b2b86e0 Generate trampoline, predicate,
and program function together") to generate the trampoline, predicate,
and clause body as a single BPF program.  This was necessary because
BPF requires a BPF program and any functions it calls to be loaded as
a single block of object code.

Finally, the predicate was integrated into the compiled clause body as
a conditional ("f56710f8 Integrate the predicate (if any) into the main
program.") to ensure global variables could be used in the predicate.

As a step towards attaching a compiled clause to multiple probes (that
each need their own trampoline) we now separate the compilation of the
clause from the trampoline generation.  After the clause has been
compiled as a function (attached to an identifier), the trampoline is
generated.  The trampoline performs its typical setup, and then makes
a function call to the clause identifier.  We depend on the linking
process introduced earlier ("3c7708ba Implement linking in of
precompiled BPF code") to combine the trampoline with the compiled
clause and any pre-compiled functions it may depend on.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>