Fix disassembly output of variable names for linked programs
The implementation of linking of BPF functions into BPF programs
("3c7708ba Implement linking in of precompiled BPF code") did not
ensure that variable information was retained in the final DIFO.
This was not an issue because there was no way to print a disassembly
listing after linking, but now we actually do care.

The linking process will now ensure that variable descriptions are
copied to the final DIFO.  For situations where multiple DIFOs are
linked together, each with their own variable table, we now also
store an instruction range for which a given variable description is
valid.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>