Move the dtrace_datadesc_t from the statement to the DIFO level
The data record description was assigned to a statement and retrieved
from there when creating enablings.  However, the actual code that will
fill in the data record is in the DIFO for the clause identifier while
the statement covers both the trampoline and the clause.

The data record description (dtrace_datadesc_t) is now stored in the
DIFO and retrieved from there.

We are still creating the enablings based on statements, so we move the
data record description from the clause to the trampoline (which is
created as a program that gets the clause linked in).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>