Defer creation of uprobes and kprobes for dtrace and fbt
The dtrace and fbt providers were creating uprobes and kprobes
unnecessarily early.  Those probes are not needed for determining
argument types since these providers do not provide typed arguments.
Hence, the uprobes and kprobes do not need to be created until and
unless we are going to attach BPF programs to them.  This avoids,
for example, the unnecessary creation of many tens of thousands of
kprobes for "dtrace -lP fbt" (once that works) or "dtrace -lv".

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>