Ensure machine state and buffer are set up before predicate eval
The predicate was being evaluated as early as possible in the BPF
function prologue for clauses.  It is however valid to use built-in
variables like epid in the predicate, and some of those variables
(e.g. epid) depend on the machine state setup to be completed.

This patch moves the predicate evaluation (and test) to be performed
after the machine state and output buffer have been set up.

A test is also included to verify that epid can be used (and has the
correct value) in the predicate.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>