Fix symbol sizes for /proc/kallsyms information
When /proc/kallmodsyms is not available, we fall back to using the
/proc/kallsyms.  That file does not contain symbol sizes, and we need
them when we add symbols to module symbol tables.

We force the size to be 1 to ensure that all symbols get added to the
symbol table.  Then, when the symbols are sorted, we set the size of
each symbol to be the difference between the next symbol address and
the current symbol address.  This will leave some symbols with 0 size
in the table, but those get cleaned up when ranges are formed.

If /proc/kallmodsyms is available, none of this is used because we will
have accurate symbol sizes.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>