Change compiler operation from per-clause to per-probe
The D compiler (and a lot of the program loading and enabling mechanism)
was entirely program or clause based.  This does not work with BPF
because a userspace process (like dtrace) cannot attach more than one
BPF program to any given perf event.  This means that if there are two
D clauses to be attached to the same probe, they somehow must be
combined.

The dtrace_program_strcompile() and dtrace_program_fcompile() functions
are called from the frontend (dtrace) to compile clauses into BPF
functions.

The dtrace_program_exec() function constructs a list of probes that will
get some BPF program attached to them, and constructs a list of clauses
(BPF functions) per probe.

The dtrace_go() function loops over the list of probes that need to have
a BPF program attached to them.  For each probe, a BPF trampoline program
is created that contains a sequence of function callsto BPF functions,
one each for every clause associated with that probe.  The linker is then
invoked to construct the complete program by recursively pulling in every
BPF function (compiled for clauses or precompiled) that is required for
the program.

Finally, each probe is instantiated on the system and has its BPF program
loaded and attached to it.``

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>