Fix tst.dtrace_cleanup.sh to handle delayed probe creation
Recent changes result in uprobes and kprobes being created when the
probe programs are loaded rather than during the populate phase.
This means that the string the test was waiting for (/matched [0-9]
probes/) may not be sufficient indication that the probes were
created.  So instead, we look for /:BEGIN/ which signals that the
BEGIN probe has fired.  This means that all probes were created.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>