pid: don't corrupt the names of pid modules
We compose these by doing an snprintf() to get a length, then
allocating a buffer and snprintfing into it.  Unfortunately the return
value of snprintf is the number of *characters* needed, but we use it
as if it were the number of bytes needed: so the last byte is always
chopped off.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>