pid: do not double-free the fasttrap_probe_spec_t
Like everything else in dt_pid, dt_pid_per_sym calls down to
dt_pid_error on error.  One of the things this can do for its callers is
free the fasttrap_probe_spec_t.  But the dt_pid_per_sym *allocates* this
in the first place, and always frees it on exit, including on error
paths (and always has). So freeing it in dt_pid_error as well is just
asking for trouble: and with the addition of the dynamically-allocated
module name, that trouble came home to roost, since we obviously have to
dereference the fasttrap_probe_spec_t (which has just been freed) to get
at its fields in order to free them...

Ditch the ability to free the fasttrap_probe_spec_t in dt_pid_error,
curing the problem without negative consequences.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>