Raise the locked-memory limit
BPF has needed locked memory.  On some systems, the limit may be rather
small, making DTrace rather unusable.

Check the limit, and if it is smaller than some "reasonably chosen"
limit, bump it up.

Note that it is hard for us to judge how much will be needed for some
D program.  One option is to raise the limit as needed whenever BPF
operations are performed (maps created, programs loaded) and fail,
retrying the operation, but there are multiple places in the code where
such retries would have to be introduced.

Another approach, when an operation fails, is simply to generate a
message that suggests that a higher limit (of unspecified amount) might
help.  Some versions of libbpf take this approach.

Mostly, it seems like recent patches could make the issue moot.  See
https://lkml.org/lkml/2020/7/24/1318
https://lkml.org/lkml/2020/7/24/1290

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>