Fix tst.dtrace_cleanup.sh on OL7
The test waited for a DTrace job to start up by monitoring
its output file for the ":BEGIN" message:
        tail -F $out | awk '/:BEGIN/ { exit; }'

The problem was that on OL7, this pipe would never terminate, even
if the ":BEGIN" message was seen.  Thus, the test would time out.

Nick points out this bug is fixed in coreutils 8.28 (OL7 is 8.22):

  ** Changes in behavior
  tail -f now exits immediately if the output is piped and the reader of
  the pipe terminates.  That allows `tail -f file | grep -q foo` to return
  responsively, but does make `tail -f file | :` exit immediately without
  waiting for data.  Instead one should now `tail -f file | grep -q .`

There is a related bug fixed in 8.32:

  'tail -f file | filter' no longer goes into an infinite loop
  if filter exits and SIGPIPE is ignored.
  [bug introduced in coreutils-8.28]

Change to a different mechanism of monitoring the file.  Have
it time out after a few seconds rather than waiting for the overall
test timeout.  For better diagnostics, if the monitoring fails to
detect the job start up, report what is in the output file.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>