Add pre-release banner UNLESS in test suite
A previous commit added a pre-release banner.  The commit message
said testing was decommissioned, but that was not the case; there
were simply widespread failures.  A later commit instituted some
mechanisms to filter out the banner, but they did not cover every
case, resulting in at least one known test FAIL.

Suppress the banner when the test suite is being run.  Remove the
filtering mechanisms.

Orabug: 31694125
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Nick Alcock <nick.alcock@oracle.com>