Mark test as XFAIL because it fails quite often
This test fails intermittently due to a bug with exit()
duplicating the last line of output.  Apparently, the test
passed in some round of testing and so its XFAIL was removed.
This removal is out of line with other profile-n and tick-n
tests, however, and the test fails more often than not.
So, restore the tag.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>