Improvements to the printing of annotations in disassembly listings
Annotations in disassembly listings were placed using tabs which would
sometimes cause them to not quite line up with other annotations due to
the instruction string varying in length.  This patch ensures that the
annotations all start in the same column (unless the instruction string
goes beyond the start of the annotations).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>