Add explicit dependency on sched.d to procfs.d.in
The procfs.d D library makes use of types that are defined in the
sched.d D library.  For unknown reasons, sched.d seems to get
processed before procfs.d on mosy systems.  But on those where it
does not, the processorid_t type was reported as unknown.

We now add the explciit dependency #pragma that should have been
there all along.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Suggested-by: Nick Alcock <nick.alcock@oracle.com>