Ensure that BEGIN is the first probe and END the last one
DTrace semantics require that the BEGIN probe is the first one to be
recorded in the trace buffers, and that the END probe is the last one
to be recorded in the trace buffers.  This is implemented using the
activity member of the DTrace state that is shared using a BPF map.

The default activity state is INACTIVE which means that no probe will
processed except for the BEGIN probe.  The BEGIN probe will only be
processed when the state is INACTIVE.

Once the BEGIN probe processing has completed, the activity state is
advanced to ACTIVE.  All probes other than BEGIN require the activity
state to be ACTIVE in order for them to be processed.

When the END probe has been processed, the activity state is advanced
to DRAINING.

The BEGIN and END probes are therefore always added to the list of
enablings, even if no clauses are specified for them.  In other words,
the BEGIN and END probes will always be enabled to ensure that the
activity state is updated correctly.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>