The exit() action must set the activity state to DRAINING
Now that activity state is supported, the exit() action needs to move
the activity state to DRAINING to prevent further probe firings from
being processed.

Note that this temporarily introduces some regressions because other
code is not processing the DRAINING state yet.  A future patch will
implement the use of the DRAINING state in the BEGIN and END probes,
and in a fix for the buffer processing semantics.

The recording of the exit code is now done using the dt_cg_store_val()
function since that is the preferred manner of storing values into the
output buffer.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>