Fix value of post-decrement expressions
An earlier commit ("17b0da163 Fix value of post-increment expressions")
purported to fix postfix expressions, but it formed (1 op arg) rather than
(arg op 1), as is evident not only in the code but even in the commit
message.  Arguably, the distinction is unimportant for increment, since
addition is commutative, but the patch produced incorrect results for
decrement expressions.

Fix the generated code and remove the XFAIL on the associated test.

Orabug: 31787365
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>