Optimize dt_cg_typecast() since shift amount is known
The typecast consists of shifting left and then right.  Since the shift
amount is known at the time of code generation, some optimizations are
possible.

Specifically, use IMM instructions to shift, thereby saving a register
and the instruction needed to load it.

Further, if the shift amount is zero, generate no shift instructions
at all.  This change happens to eliminate the BPF error encountered in
test/unittest/variables/bvar/tst.curcpu.d so that it now passes.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>